Add files via upload
In this section, we have prepared a free download of the project for removing the background of video images by improving the Gaussian mixture model (GMM algorithm) with C Plus language, which is based on the article Improved Adaptive Gaussian Mixture Model for Background Subtraction. In the following, some articles and explanations about the Gaussian mixture model are presented, and the free source code of this image processing and machine vision project is included with the basic article.

Image background reduction is a common task in computer vision. In this paper we analyze the usual pixel level approach. In this paper, an efficient adaptive algorithm is developed using Gaussian mixture probability density. Recursive equations are used to continuously update the parameters as well as to simultaneously select the appropriate number of components for each pixel.

The Gaussian Mixture Model is one of the relatively new methods that have been proposed for image classification. The use of this GMM algorithm is increasing due to satisfactory results and relatively high speed. Gaussian mixture model is used to a significant extent in image density modeling and clustering. These models have full estimation capability, because they can accurately model any type of probability density function (PDF) with a sufficient number of Gaussian functions.